osint_backend/requirements_utf8.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
alembic==1.15.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
asyncio==4.0.0
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
boto3==1.42.68
botocore==1.42.68
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
contourpy==1.3.1
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
Django==4.2.11
django-cors-headers==4.3.1
django-jazzmin==3.0.3
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
docutils==0.21.2
elevenlabs==2.39.0
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.135.1
fastjsonschema==2.21.1
filelock==3.25.2
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.1.1
gTTS==2.5.4
gunicorn==21.2.0
h11==0.14.0
hf-xet==1.4.2
holehe==1.61
html2text==2024.2.26
httpcore==1.0.7
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.10
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.1.0
joblib==1.5.3
jq==1.11.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lab==8.3
lxml==5.3.2
lxml_html_clean==0.4.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.1
mysql-connector-python==9.2.0
mysqlclient==2.2.7
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
node==1.2.2
notebook==7.3.2
notebook_shim==0.2.4
num2words==0.5.14
numpy==2.2.1
odict==1.9.0
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pdf2image==1.17.0
phonenumbers==9.0.25
pillow==11.1.0
platformdirs==4.3.6
plumber==1.7
polib==1.2.0
proglog==0.1.12
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.4.1
psutil==6.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.11.0
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypdf==5.1.0
PyPDF2==3.0.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-json-logger==3.2.1
pyttsx3==2.98
pytube==15.0.0
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2026.2.28
reportlab==4.3.1
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.3.3
rjsmin==1.2.4
rpds-py==0.22.3
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.3.0
setuptools==82.0.1
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
termcolor==3.3.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.10.0
tornado==6.4.2
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
trio==0.33.0
txt2tags==3.9
typer==0.24.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.42.0
vobject==0.9.9
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
Werkzeug==3.1.3
wheel==0.46.3
whisper==1.1.10
widgetsnbextension==4.0.13
xlrd==2.0.1
yarl==1.23.0
zeep==4.3.1
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.1
zope.event==5.0
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.proxy==6.1
